backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
openai==1.12.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
stripe==7.11.0
boto3==1.34.34
requests==2.31.0 